LayoutTests/platform/win/accessibility/text-role.html

<html>
    <head>
        <script src="../../../fast/js/resources/js-test-pre.js"></script>

        <script>
            if (window.testRunner)
                testRunner.dumpAsText();
        </script>
    </head>
    <body id="body">
        <p>
            This test verifies that, on Windows, text nodes return "editable
            text" for their role value.
            <br><br>
            <a id="link" href="https://bugs.webkit.org/show_bug.cgi?id=32631">Bug 32631 -
            MSAA: Accessibility role of text nodes is wrong</a>
        </p>

        <p id="console"></p>

        <script>
            if (window.accessibilityController) {
                document.getElementById("link").focus();

                var e = accessibilityController.focusedElement.childAtIndex(0);
                shouldBe('e.role', '"editable text"');
            }
        </script>
    </body>
</html>
